Moved LinuxAppIconExtractor into Linux folder
