+++ src/main/Core/ImageGenerator/Linux/index.ts
@@ -1,0 +1,1 @@
+export * from "./LinuxAppIconExtractor";
